move all string parsing code from renderText into the utils file
